Adição dos forms para CRUD de produtos e categoria

+++ estoque/models.py
@@ -10,8 +10,8 @@ class Produto(models.Model):
     nome_produto = models.CharField(max_length=50)
     categoria = models.ForeignKey(Categoria, on_delete=models.SET_NULL, null=True)
     quantidade_estoque = models.PositiveIntegerField(default = 0)
-    valor_custo = models.FloatField(default=0)
-    valor_venda = models.FloatField(default=0)
+    valor_custo = models.DecimalField(decimal_places=2, max_digits=10, default=0)
+    valor_venda = models.DecimalField(decimal_places=2, max_digits=10, default=0)
     disponivel = models.BooleanField(default=False)
     def __str__(self):
         return self.nome_produto
